+++ database/migrations/2022_10_15_170247_create_event_photos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('event_photos', function (Blueprint $table) {
             $table->id();
-            $table->foreignId('event_id')->referenced('events')->cascadeOnDelete();
+            $table->foreignId('event_id')->referenced('events')->cascadeOnDelete()->cascadeOnUpdate();
             $table->string('path');
             $table->timestamps();
         });
@@ -28,6 +28,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('event_photos');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('event_photos');
+        }
     }
 };
