+++ database/migrations/2022_10_15_170611_create_orders_table.php
@@ -31,6 +31,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('orders');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('orders');
+        }
     }
 };
