+++ database/migrations/2022_10_15_181151_create_feedbacks_table.php
@@ -29,6 +29,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('feedbacks');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('feedbacks');
+        }
     }
 };
